refactor: separate NFC and Strava page
